[NET-765] Rename database manager file
