+++ cmdb/framework/cmdb_base.py
@@ -18,7 +18,7 @@ from abc import ABC
 from typing import List
 
 from cmdb.utils.error import CMDBError
-from cmdb.database.database_manager import DatabaseManagerMongo
+from cmdb.database.managers import DatabaseManagerMongo
 
 LOGGER = logging.getLogger(__name__)
