+++ cmdb/interface/net_app/__init__.py
@@ -31,7 +31,7 @@ system_config_reader = SystemConfigReader()
 
 def create_app(event_queue):
     # Create managers
-    from cmdb.database.database_manager import DatabaseManagerMongo
+    from cmdb.database.managers import DatabaseManagerMongo
     app_database = DatabaseManagerMongo(
         **system_config_reader.get_all_values_from_section('Database')
     )
