+++ cmdb/user_management/managers/account_manager.py
@@ -15,7 +15,7 @@
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 from typing import Union, List
 
-from cmdb.database.database_manager import DatabaseManagerMongo
+from cmdb.database.managers import DatabaseManagerMongo
 from cmdb.framework.results import IterationResult
 from cmdb.framework.utils import Collection, PublicID
 from cmdb.manager import ManagerBase
