+++ cmdb/user_management/user_manager.py
@@ -17,7 +17,7 @@
 import logging
 from typing import List
 
-from cmdb.database.database_manager import DatabaseManagerMongo
+from cmdb.database.managers import DatabaseManagerMongo
 from cmdb.database.errors.database_errors import NoDocumentFound
 from cmdb.framework.cmdb_base import CmdbManagerBase, ManagerGetError, ManagerInsertError, ManagerUpdateError, \
     ManagerDeleteError
